@@ -19,6 +19,7 @@
   },
   "contributors": [],
   "license": "MIT",
+  "type": "module",
   "main": "bin/pwned.js",
   "bin": "bin/pwned.js",
   "directories": {
@@ -61,7 +62,7 @@
   },
   "homepage": "https://wkovacs64.github.io/pwned",
   "engines": {
-    "node": ">= 10"
+    "node": ">= 12.20"
   },
   "dependencies": {
     "common-tags": "^1.8.0",
@@ -80,6 +81,7 @@
     "@babel/preset-typescript": "7.13.0",
     "@commitlint/cli": "12.1.1",
     "@commitlint/config-conventional": "12.1.1",
+    "@jest/globals": "26.6.2",
     "@types/common-tags": "1.8.0",
     "@types/jest": "26.0.23",
     "@types/node": "12.20.11",
@@ -88,6 +90,8 @@
     "@types/yargs": "16.0.1",
     "@wkovacs64/prettier-config": "3.0.0",
     "babel-jest": "26.6.3",
+    "babel-plugin-add-import-extension": "1.5.1",
+    "babel-plugin-transform-import-meta": "2.0.0",
     "codecov": "3.8.1",
     "commitizen": "4.2.3",
     "cross-env": "7.0.3",
