@@ -1,3 +1,3 @@
-import * as pkg from '../../package.json';
+import { pkg } from './pkg';
 
 export const userAgent = `${pkg.name} ${pkg.version}`;
